# backend/bait-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API in Laravel",
        "description": "API documentation with Swagger in Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/states": {
            "get": {
                "tags": [
                    "User States"
                ],
                "summary": "Get all user states",
                "description": "Returns a list of all user states (e.g. active, suspended).",
                "operationId": "getAllStates",
                "responses": {
                    "200": {
                        "description": "List of user states",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserStateSchema"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User States"
                ],
                "summary": "Create a new user state",
                "description": "Creates a new user state in the system.",
                "operationId": "e0794cb4901b41e4baa96d64e2f3de62",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "muted"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "State created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserStateSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/states/{state}": {
            "put": {
                "tags": [
                    "User States"
                ],
                "summary": "Update a user state",
                "description": "Updates an existing user state.",
                "operationId": "85e9e16c11ce4be0c5b13b14383fb785",
                "parameters": [
                    {
                        "name": "state",
                        "in": "path",
                        "description": "ID of the state to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 4
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "banned"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "State updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserStateSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated"
                    },
                    "404": {
                        "description": "State not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User States"
                ],
                "summary": "Delete a user state",
                "description": "Deletes an existing user state.",
                "operationId": "f84233cc731f4228509fa8c06a17bde2",
                "parameters": [
                    {
                        "name": "state",
                        "in": "path",
                        "description": "ID of the state to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 4
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "State deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated"
                    },
                    "404": {
                        "description": "State not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chats": {
            "get": {
                "tags": [
                    "Chats"
                ],
                "summary": "List all chats for the authenticated user",
                "description": "Retrieves a list of chats where the authenticated user is a participant.",
                "operationId": "getUserChats",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChatSchema"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "Create a new chat",
                "description": "Creates a new chat and adds the authenticated user and specified participants.",
                "operationId": "createChat",
                "requestBody": {
                    "description": "Array of participant user IDs to include in the chat.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "participants"
                                ],
                                "properties": {
                                    "participants": {
                                        "description": "An array of user IDs.",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chat created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chats/{chat}": {
            "get": {
                "tags": [
                    "Chats"
                ],
                "summary": "Get details of a specific chat",
                "description": "Retrieves the details of a single chat, including participants and messages. Authenticated user must be a participant.",
                "operationId": "getChatById",
                "parameters": [
                    {
                        "name": "chat",
                        "in": "path",
                        "description": "The ID of the chat.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - User is not a participant of this chat"
                    },
                    "404": {
                        "description": "Chat not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AvatarSchema": {
                "title": "Avatar Schema",
                "required": [
                    "id",
                    "url_avatars"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "url_avatars": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/storage/avatars/avatar1.jpg"
                    }
                },
                "type": "object"
            },
            "BannerSchema": {
                "title": "Banner Schema",
                "required": [
                    "id",
                    "url_banners"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "url_banners": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/storage/banners/banner1.jpg"
                    }
                },
                "type": "object"
            },
            "ChatSchema": {
                "title": "Chat Schema",
                "description": "Represents a chat conversation between users.",
                "properties": {
                    "id": {
                        "description": "The chat's unique identifier.",
                        "type": "integer",
                        "example": 1
                    },
                    "participants": {
                        "description": "List of users participating in the chat.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserSchema"
                        }
                    },
                    "last_message": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/MessageSchema"
                            }
                        ],
                        "nullable": true,
                        "description": "The most recent message in the chat."
                    },
                    "created_at": {
                        "description": "Timestamp when the chat was created.",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-05T10:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "MessageSchema": {
                "title": "Message Schema",
                "description": "Represents a message within a chat.",
                "properties": {
                    "id": {
                        "description": "The message's unique identifier.",
                        "type": "integer",
                        "example": 101
                    },
                    "content_messages": {
                        "description": "The text content of the message.",
                        "type": "string",
                        "example": "Hello, how are you?"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserSchema"
                    },
                    "created_at": {
                        "description": "Timestamp when the message was sent.",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-05T12:30:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "UserSchema": {
                "title": "User Schema",
                "required": [
                    "id",
                    "username",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 123
                    },
                    "username": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "role": {
                        "description": "User's role name",
                        "type": "string",
                        "example": "admin",
                        "nullable": true
                    },
                    "state": {
                        "description": "User's state name",
                        "type": "string",
                        "example": "active",
                        "nullable": true
                    },
                    "avatar": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/AvatarSchema"
                            }
                        ],
                        "nullable": true
                    },
                    "banner": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/BannerSchema"
                            }
                        ],
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-05T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-06T12:34:56Z"
                    }
                },
                "type": "object"
            },
            "UserStateSchema": {
                "title": "User State Schema",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "active"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a JWT token to authenticate",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}